+++ TCC.Core/SessionManager.cs
@@ -122,7 +122,7 @@ namespace TCC
             }
         }
 
-        public static void SetPlayerMaxHP(ulong target, long maxHP)
+        public static void SetPlayerMaxHP(ulong target, int maxHP)
         {
             if (target == CurrentPlayer.EntityId)
             {
